entity package refactoring
