+++ spring-data-jpa/src/main/java/com/nubiform/springdatajpa/repository/OwnerRepository.java
@@ -1,6 +1,6 @@
 package com.nubiform.springdatajpa.repository;
 
-import com.nubiform.springdatajpa.model.Owner;
+import com.nubiform.springdatajpa.repository.entity.Owner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OwnerRepository extends JpaRepository<Owner, Integer> {
