+++ spring-data-jpa/src/main/java/com/nubiform/springdatajpa/repository/entity/Owner.java
@@ -1,4 +1,4 @@
-package com.nubiform.springdatajpa.model;
+package com.nubiform.springdatajpa.repository.entity;
 
 import lombok.Builder;
 import lombok.Getter;
